chore(common): cleanup and reorganize hextobin to remove fs dependency
fs dependency removed from the main hextobin function, only in the
filesystem.ts version. Also supports loading a segment of a hex file
for isolated tests.

Test-bot: skip
