+++ common/tools/hextobin/hextobin.ts
@@ -1,6 +1,13 @@
 #!/usr/bin/env node
+/*
+ * Keyman is copyright (C) SIL Global. MIT License.
+ *
+ * hextobin - a tool to convert a structured hex text file format into a binary file.
+ * This is the command line wrapper.
+ */
+
 import { Command } from 'commander';
-import hextobin from './index.js';
+import { hextobinFromFile } from './filesystem.js';
 
 let inputFilename: string = "";
 let outputFilename: string = "";
@@ -43,16 +50,15 @@ function exitDueToUsageError(message: string): never  {
   return process.exit(64); // SysExits.EX_USAGE
 }
 
-hextobin(inputFilename, outputFilename, { silent: false })
-  .then((buffer) => {
-    if (buffer) {
-      process.exit(0); // OK
-    } else {
-      console.error(`${program._name}: Failed.`);
-      process.exit(1); // no buffer - some err.
-    }
-  }, (err) => {
-    console.error(err);
+try {
+  if(hextobinFromFile(inputFilename, outputFilename, { silent: false })) {
+    process.exit(0); // OK
+  } else {
     console.error(`${program._name}: Failed.`);
-    process.exit(1);
-  });
+    process.exit(1); // no buffer - some err.
+  }
+} catch(err) {
+  console.error(err);
+  console.error(`${program._name}: Failed.`);
+  process.exit(1);
+}
